Initial commit with basic apps for the template
Often times I want to start new ideas, but just the thought of having to
configure Rails from the start is a major blocker for me. It aggravates the
cases where I need more than just Rails.

This commit initializes a monorepo template repository that I can reuse
overtime. It introduces 3 directories with barebone apps

- app: this is a Rails app for users to access via `rails server`.
- analysis: this is a typescript codebase for whatever you need it to (reasoning
  below).
- analysis-node: this is just a thin `node` wrapper around `analysis`. You can't
  interact with `analysis` directly from the Rails app, you have to use the
  wrapper via stdout.

The advantage of a pure Typescript app is that it can be built as JS and
embedded into mobile apps. That means that if you have an app that its core
domain features can be written in TS, you can also run it within iOS, Android,
in the server and more.

The iOS, Android apps can focus on UI while calling the JS code for core
functionality. For example, you could create a crawler in TS, and have it run in
the iOS with the same rules as on the server, distributing the load.

I used this architecture for a chess app. The analysis part is heavy, and so
when it's run in iOS, the entire analysis logic runs locally. When running on
the server, it runs with the same logic.

On the other hand, if it was built with Rails only, you'd have to rewrite it for
iOS and then again for Android.